S.D.Process: throw PNSE on HWND depending funcs
