+++ src/System.Diagnostics.Process/src/System/Diagnostics/Process.cs
@@ -65,13 +65,6 @@ namespace System.Diagnostics
         private StreamReader _standardError;
         private bool _disposed;
 
-        private bool _haveMainWindow;
-        private IntPtr _mainWindowHandle;
-        private string _mainWindowTitle;
-
-        private bool _haveResponding;
-        private bool _responding;
-
         private static object s_createProcessLock = new object();
 
         private StreamReadMode _outputStreamReadMode;
@@ -812,33 +805,6 @@ namespace System.Diagnostics
             }
         }
 
-        public bool Responding
-        {
-            get
-            {
-                if (!_haveResponding)
-                {
-                    _responding = IsRespondingCore();
-                    _haveResponding = true;
-                }
-
-                return _responding;
-            }
-        }
-
-        public string MainWindowTitle
-        {
-            get
-            {
-                if (_mainWindowTitle == null)
-                {
-                    _mainWindowTitle = GetMainWindowTitle();
-                }
-
-                return _mainWindowTitle;
-            }
-        }
-
         public bool CloseMainWindow()
         {
             return CloseMainWindowCore();
@@ -852,21 +818,6 @@ namespace System.Diagnostics
         public bool WaitForInputIdle(int milliseconds)
         {
             return WaitForInputIdleCore(milliseconds);
-        }
-
-        public IntPtr MainWindowHandle
-        {
-            get
-            {
-                if (!_haveMainWindow)
-                {
-                    EnsureState(State.IsLocal | State.HaveId);
-                    _mainWindowHandle = ProcessManager.GetMainWindowHandle(_processId);
-
-                    _haveMainWindow = true;
-                }
-                return _mainWindowHandle;
-            }
         }
 
         public ISynchronizeInvoke SynchronizingObject { get; set; }
