+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessManager.WinRT.cs
@@ -1,0 +1,21 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Globalization;
+using System.Runtime.InteropServices;
+using System.Threading;
+using Microsoft.Win32.SafeHandles;
+
+namespace System.Diagnostics
+{
+    internal static partial class ProcessManager
+    {
+        public static IntPtr GetMainWindowHandle(int processId)
+        {
+            throw new PlatformNotSupportedException();
+        }
+    }
+}
