+++ src/System.Diagnostics.Process/tests/ProcessWaitingTests.cs
@@ -217,6 +217,7 @@ namespace System.Diagnostics.Tests
         }
 
         [Fact]
+        [SkipOnTargetFramework(TargetFrameworkMonikers.Uap, "HWND not available")]
         public void WaitForInputIdle_NotDirected_ThrowsInvalidOperationException()
         {
             var process = new Process();
